@@ -12,24 +12,22 @@
     "require-dev": {
         "adlawson/vfs": "^0.12.1",
         "behat/behat": "^3.10",
+        "behat/mink-selenium2-driver": "^1.5",
         "friends-of-behat/mink": "^1.9",
         "friends-of-behat/mink-extension": "^2.5",
-        "behat/mink-selenium2-driver": "^1.4",
         "centreon/centreon-test-lib": "dev-password-security-policy",
-        "phpstan/phpstan": "^0.12.59",
-        "phpstan/phpstan-beberlei-assert": "^0.12",
+        "phpstan/phpstan": "^1.3.0",
+        "phpstan/phpstan-beberlei-assert": "^1.0.0",
         "phpunit/phpunit": "^9.5",
-        "sensiolabs/security-checker": "^6.0",
-        "squizlabs/php_codesniffer": "^3.5",
-        "symfony/phpunit-bridge": "4.4.*",
-        "symfony/stopwatch": "^4.4",
-        "symfony/twig-bundle": "^4.4",
-        "symfony/var-dumper": "4.4.*",
-        "symfony/web-profiler-bundle": "^4.4",
-        "twig/twig": "^2.0",
+        "squizlabs/php_codesniffer": "3.6.1",
+        "symfony/phpunit-bridge": "6.0.*",
+        "symfony/stopwatch": "^5.4",
+        "symfony/twig-bundle": "^5.4",
+        "symfony/var-dumper": "5.4.*",
+        "symfony/web-profiler-bundle": "^5.4",
+        "twig/twig": "^3.3",
         "webmozart/assert": "^1.8",
-        "zircote/swagger-php": "^3.0",
-        "psr/container": "~1.0.0"
+        "zircote/swagger-php": "^4.0"
     },
     "require": {
         "beberlei/assert": "^3.3",
@@ -42,36 +40,38 @@
         "ext-json": "*",
         "ext-openssl": "*",
         "ext-phar": "*",
-        "friendsofsymfony/rest-bundle": "^2.5",
-        "jms/serializer-bundle": "^3.9",
+        "friendsofsymfony/rest-bundle": "^3.0",
+        "jms/serializer-bundle": "^4.0",
         "justinrainbow/json-schema": "^5.2",
         "nelmio/cors-bundle": "^2.1",
         "openpsa/quickform": "3.3.*",
         "pear/pear-core-minimal": "^1.10",
-        "php": ">=7.2.24",
         "phpdocumentor/reflection-docblock": "^5.2",
         "pimple/pimple": "^3.2",
-        "sensio/framework-extra-bundle": "^5.3",
+        "sensio/framework-extra-bundle": "^6.2",
         "smarty/smarty": "^3.1",
         "smarty-gettext/smarty-gettext": "^1.6",
-        "symfony/console": "4.4.*",
-        "symfony/dotenv": "4.4.*",
-        "symfony/expression-language": "4.4.*",
-        "symfony/filesystem": "4.4.*",
-        "symfony/finder": "4.4.*",
-        "symfony/flex": "^1.9",
-        "symfony/framework-bundle": "4.4.*",
-        "symfony/http-client": "4.4.*",
+        "symfony/config": "5.4.*",
+        "symfony/console": "5.4.*",
+        "symfony/dotenv": "5.4.*",
+        "symfony/expression-language": "5.4.*",
+        "symfony/filesystem": "5.4.*",
+        "symfony/finder": "5.4.*",
+        "symfony/flex": "^1.17",
+        "symfony/framework-bundle": "5.4.*",
+        "symfony/http-client": "5.4.*",
+        "symfony/http-kernel": "5.4.*",
         "symfony/maker-bundle": "^1.11",
         "symfony/monolog-bundle": "^3.7",
-        "symfony/options-resolver": "4.4.*",
-        "symfony/property-access": "4.4.*",
-        "symfony/property-info": "4.4.*",
-        "symfony/security-bundle": "4.4.*",
-        "symfony/serializer": "4.4.*",
-        "symfony/translation": "4.4.*",
-        "symfony/validator": "4.4.*",
-        "symfony/yaml": "4.4.*"
+        "symfony/options-resolver": "5.4.*",
+        "symfony/property-access": "5.4.*",
+        "symfony/property-info": "5.4.*",
+        "symfony/security-bundle": "5.4.*",
+        "symfony/serializer": "5.4.*",
+        "symfony/translation": "5.4.*",
+        "symfony/validator": "5.4.*",
+        "symfony/yaml": "5.4.*",
+        "psr/container": "1.1.1"
     },
     "autoload": {
         "psr-4": {
