@@ -38,17 +38,20 @@ abstract class AbstractController extends AbstractFOSRestController
 
     public function denyAccessUnlessGrantedForApiConfiguration(): void
     {
-        parent::denyAccessUnlessGranted([
+        parent::denyAccessUnlessGranted(
             static::ROLE_API_CONFIGURATION,
+            null,
             static::ROLE_API_CONFIGURATION_EXCEPTION_MESSAGE
-        ]);
+        );
     }
 
     public function denyAccessUnlessGrantedForApiRealtime(): void
     {
-        parent::denyAccessUnlessGranted([
+        parent::denyAccessUnlessGranted(
             static::ROLE_API_REALTIME,
-        ], static::ROLE_API_REALTIME_EXCEPTION_MESSAGE);
+            null,
+            static::ROLE_API_REALTIME_EXCEPTION_MESSAGE
+        );
     }
 
     /**
