@@ -22,17 +22,18 @@ declare(strict_types=1);
 
 namespace Centreon\Application\Controller\Configuration;
 
-use Centreon\Domain\Contact\Interfaces\ContactInterface;
-use Centreon\Domain\Entity\EntityValidator;
-use Centreon\Domain\Proxy\Interfaces\ProxyServiceInterface;
-use Centreon\Domain\Proxy\Proxy;
-use Centreon\Application\Controller\AbstractController;
 use FOS\RestBundle\View\View;
-use JMS\Serializer\Exception\ValidationFailedException;
+use Centreon\Domain\Proxy\Proxy;
+use Centreon\Domain\Contact\Contact;
 use JMS\Serializer\SerializerInterface;
+use Centreon\Domain\Entity\EntityValidator;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Centreon\Application\Controller\AbstractController;
+use JMS\Serializer\Exception\ValidationFailedException;
+use Centreon\Domain\Contact\Interfaces\ContactInterface;
 use Symfony\Component\HttpKernel\Exception\HttpException;
+use Centreon\Domain\Proxy\Interfaces\ProxyServiceInterface;
 
 /**
  * This class is design to manage all API REST requests concerning the proxy configuration.
@@ -64,7 +65,11 @@ class ProxyController extends AbstractController
     {
         $this->denyAccessUnlessGrantedForApiConfiguration();
 
-        if (!$this->getUser()->isAdmin() && !$this->isGranted('ROLE_ADMINISTRATION_PARAMETERS_CENTREON_UI_RW')) {
+        /**
+         * @var Contact $contact
+         */
+        $contact = $this->getUser();
+        if (!$contact->isAdmin() && !$this->isGranted('ROLE_ADMINISTRATION_PARAMETERS_CENTREON_UI_RW')) {
             return $this->view(null, Response::HTTP_FORBIDDEN);
         }
         return $this->view($this->proxyService->getProxy());
