@@ -99,8 +99,13 @@ class PlatformTopologyController extends AbstractController
         // check user rights
         $this->denyAccessUnlessGrantedForApiConfiguration();
 
+        /**
+         * @var Contact $contact
+         */
+        $contact = $this->getUser();
+
         // Check Topology access to Configuration > Pollers page
-        if (!$this->getUser()->hasTopologyRole(Contact::ROLE_CONFIGURATION_MONITORING_SERVER_READ_WRITE)) {
+        if (!$contact->hasTopologyRole(Contact::ROLE_CONFIGURATION_MONITORING_SERVER_READ_WRITE)) {
             return $this->view(null, Response::HTTP_FORBIDDEN);
         }
 
@@ -155,6 +160,10 @@ class PlatformTopologyController extends AbstractController
         $this->denyAccessUnlessGrantedForApiConfiguration();
 
         // Check Topology access to Configuration > Pollers page
+
+        /**
+         * @var Contact $user
+         */
         $user = $this->getUser();
         if (
             !$user->hasTopologyRole(Contact::ROLE_CONFIGURATION_MONITORING_SERVER_READ)
@@ -207,7 +216,11 @@ class PlatformTopologyController extends AbstractController
         $this->denyAccessUnlessGrantedForApiConfiguration();
 
         // Check Topology access to Configuration > Pollers page
-        if (!$this->getUser()->hasTopologyRole(Contact::ROLE_CONFIGURATION_MONITORING_SERVER_READ_WRITE)) {
+        /**
+         * @var Contact $contact
+         */
+        $contact = $this->getUser();
+        if (!$contact->hasTopologyRole(Contact::ROLE_CONFIGURATION_MONITORING_SERVER_READ_WRITE)) {
             return $this->view(null, Response::HTTP_FORBIDDEN);
         }
 
