@@ -224,8 +224,8 @@ class NavigationList implements JsonSerializable
     /**
      * Extract the array without keys to avoid serialization into objects
      *
-     * @param  array<array> $naviList
-     * @return array<array>
+     * @param  array<mixed> $naviList
+     * @return array<mixed>
      */
     private function removeKeysFromArray(array $naviList): array
     {
