@@ -75,9 +75,8 @@ final class ContactRepositoryRDB implements ContactRepositoryInterface
 
         if (($result = $statement->fetch(\PDO::FETCH_ASSOC)) !== false) {
             $contact = $this->createContact($result);
-            $this->addActionRules($contact);
-            $this->addTopologyRules($contact);
         }
+
         return $contact;
     }
 
@@ -107,8 +106,6 @@ final class ContactRepositoryRDB implements ContactRepositoryInterface
         $contact = null;
         if (($result = $statement->fetch(\PDO::FETCH_ASSOC)) !== false) {
             $contact = $this->createContact($result);
-            $this->addActionRules($contact);
-            $this->addTopologyRules($contact);
         }
 
         return $contact;
@@ -141,8 +138,6 @@ final class ContactRepositoryRDB implements ContactRepositoryInterface
         $contact = null;
         if (($result = $statement->fetch(\PDO::FETCH_ASSOC)) !== false) {
             $contact = $this->createContact($result);
-            $this->addActionRules($contact);
-            $this->addTopologyRules($contact);
         }
 
         return $contact;
@@ -174,11 +169,12 @@ final class ContactRepositoryRDB implements ContactRepositoryInterface
         $statement->bindValue(':token', $token, \PDO::PARAM_STR);
         $statement->execute();
 
+        $contact = null;
         if ($result = $statement->fetch(\PDO::FETCH_ASSOC)) {
-            return $this->createContact($result);
+            $contact = $this->createContact($result);
         }
 
-        return null;
+        return $contact;
     }
 
     /**
@@ -380,7 +376,7 @@ final class ContactRepositoryRDB implements ContactRepositoryInterface
             }
         }
 
-        return (new Contact())
+        $contact = (new Contact())
             ->setId((int) $contact['contact_id'])
             ->setName($contact['contact_name'])
             ->setAlias($contact['contact_alias'])
@@ -398,6 +394,11 @@ final class ContactRepositoryRDB implements ContactRepositoryInterface
             ->setDefaultPage($page)
             ->setUseDeprecatedPages($contact['show_deprecated_pages'] === '1')
             ->setOneClickExportEnabled($contact['enable_one_click_export'] === '1');
+
+        $this->addActionRules($contact);
+        $this->addTopologyRules($contact);
+
+        return $contact;
     }
 
     /**
