@@ -44,11 +44,12 @@ class TimeperiodWebservice extends Webservice\WebServiceAbstract implements
 
     /**
      * {@inheritdoc}
+     * @return array<ServiceProvider::CENTREON_PAGINATION>
      */
     public static function dependencies(): array
     {
         return [
-            ServiceProvider::CENTREON_PAGINATION,
+            ServiceProvider::CENTREON_PAGINATION
         ];
     }
 
