@@ -40,6 +40,12 @@ class TimeperiodWebserviceTest extends TestCase
     protected const METHOD_GET_LIST = 'getList';
 
     /**
+     *
+     * @var TimeperiodWebservice|\PHPUnit\Framework\MockObject\MockObject
+     */
+    private $webservice;
+
+    /**
      * {@inheritdoc}
      */
     protected function setUp(): void
@@ -63,7 +69,7 @@ class TimeperiodWebserviceTest extends TestCase
     /**
      * Test the method getList
      */
-    public function testGetList()
+    public function testGetList(): void
     {
         // without applied filters
         $this->mockQuery();
@@ -72,8 +78,9 @@ class TimeperiodWebserviceTest extends TestCase
 
     /**
      * Test the method getList with different filter
+     *
      */
-    public function testGetList2()
+    public function testGetList2(): void
     {
         // with search, searchByIds, limit, and offset
         $this->mockQuery([
@@ -87,8 +94,9 @@ class TimeperiodWebserviceTest extends TestCase
 
     /**
      * Test the method getName
+     *
      */
-    public function testGetName()
+    public function testGetName(): void
     {
         $this->assertEquals('centreon_timeperiod', TimeperiodWebservice::getName());
     }
