@@ -37,9 +37,14 @@ class TimeperiodRepositoryTest extends TestCase
     use Traits\PaginationListTrait;
 
     /**
-     * @var array
+     * @var array<int, array<string, array<int, array<string, int|string>>|string>>
      */
     protected $datasets = [];
+
+    /**
+     * @var CentreonDB
+     */
+    private $db;
 
     /**
      * {@inheritdoc}
@@ -93,7 +98,7 @@ class TimeperiodRepositoryTest extends TestCase
     /**
      * Test the method entityClass
      */
-    public function testEntityClass()
+    public function testEntityClass(): void
     {
         $this->assertEquals(Timeperiod::class, TimeperiodRepository::entityClass());
     }
@@ -101,7 +106,7 @@ class TimeperiodRepositoryTest extends TestCase
     /**
      * Test the method checkListOfIds
      */
-    public function testCheckListOfIds()
+    public function testCheckListOfIds(): void
     {
         $this->checkListOfIdsTrait(
             TimeperiodRepository::class,
@@ -112,7 +117,7 @@ class TimeperiodRepositoryTest extends TestCase
     /**
      * Test the method getPaginationList
      */
-    public function testGetPaginationList()
+    public function testGetPaginationList(): void
     {
         $this->getPaginationListTrait($this->datasets[0]['data'][0]);
     }
@@ -120,7 +125,7 @@ class TimeperiodRepositoryTest extends TestCase
     /**
      * Test the method getPaginationList with different set of arguments
      */
-    public function testGetPaginationListWithArguments()
+    public function testGetPaginationListWithArguments(): void
     {
         $this->getPaginationListTrait(
             $this->datasets[1]['data'][0],
@@ -140,7 +145,7 @@ class TimeperiodRepositoryTest extends TestCase
     /**
      * Test the method getPaginationListTotal
      */
-    public function testGetPaginationListTotal()
+    public function testGetPaginationListTotal(): void
     {
         $this->getPaginationListTotalTrait(
             $this->datasets[2]['data'][0]['number']
