@@ -30,7 +30,6 @@ use Security\Domain\Authentication\Model\ProviderConfiguration;
 use Centreon\Domain\Authentication\UseCase\Authenticate;
 use Centreon\Domain\Contact\Interfaces\ContactInterface;
 use Security\Domain\Authentication\Model\AuthenticationTokens;
-use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Centreon\Domain\Authentication\UseCase\AuthenticateRequest;
 use Centreon\Domain\Contact\Interfaces\ContactServiceInterface;
 use Centreon\Domain\Authentication\UseCase\AuthenticateResponse;
@@ -43,6 +42,9 @@ use Security\Domain\Authentication\Interfaces\SessionRepositoryInterface;
 use Security\Domain\Authentication\Interfaces\ProviderServiceInterface;
 use Security\Domain\Authentication\Interfaces\AuthenticationServiceInterface;
 use Centreon\Domain\Repository\Interfaces\DataStorageEngineInterface;
+use Symfony\Component\HttpFoundation\RequestStack;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Session\SessionInterface;
 
 /**
  * @package Tests\Centreon\Domain\Authentication\UseCase
@@ -63,6 +65,16 @@ class AuthenticateTest extends TestCase
      * @var ContactServiceInterface|\PHPUnit\Framework\MockObject\MockObject
      */
     private $contactService;
+
+    /**
+     * @var RequestStack|\PHPUnit\Framework\MockObject\MockObject
+     */
+    private $requestStack;
+
+    /**
+     * @var Request|\PHPUnit\Framework\MockObject\MockObject
+     */
+    private $request;
 
     /**
      * @var SessionInterface|\PHPUnit\Framework\MockObject\MockObject
@@ -114,7 +126,6 @@ class AuthenticateTest extends TestCase
         $this->authenticationService = $this->createMock(AuthenticationServiceInterface::class);
         $this->providerService = $this->createMock(ProviderServiceInterface::class);
         $this->contactService = $this->createMock(ContactServiceInterface::class);
-        $this->session = $this->createMock(SessionInterface::class);
         $this->provider = $this->createMock(ProviderInterface::class);
         $this->contact = $this->createMock(ContactInterface::class);
         $this->authenticationTokens = $this->createMock(AuthenticationTokens::class);
@@ -123,6 +134,21 @@ class AuthenticateTest extends TestCase
         $this->authenticationRepository = $this->createMock(AuthenticationRepositoryInterface::class);
         $this->sessionRepository = $this->createMock(SessionRepositoryInterface::class);
         $this->dataStorageEngine = $this->createMock(DataStorageEngineInterface::class);
+        $this->session = $this->createMock(SessionInterface::class);
+        $this->session
+            ->expects($this->any())
+            ->method('getId')
+            ->willReturn('session_abcd');
+        $this->request = $this->createMock(Request::class);
+        $this->request
+            ->expects($this->any())
+            ->method('getSession')
+            ->willReturn($this->session);
+        $this->requestStack = $this->createMock(RequestStack::class);
+        $this->requestStack
+            ->expects($this->any())
+            ->method('getCurrentRequest')
+            ->willReturn($this->request);
     }
 
     /**
@@ -135,7 +161,7 @@ class AuthenticateTest extends TestCase
             $this->authenticationService,
             $this->providerService,
             $this->contactService,
-            $this->session,
+            $this->requestStack,
             $this->menuService,
             $this->authenticationRepository,
             $this->sessionRepository,
@@ -177,7 +203,7 @@ class AuthenticateTest extends TestCase
             $this->authenticationService,
             $this->providerService,
             $this->contactService,
-            $this->session,
+            $this->requestStack,
             $this->menuService,
             $this->authenticationRepository,
             $this->sessionRepository,
@@ -219,7 +245,7 @@ class AuthenticateTest extends TestCase
             $this->authenticationService,
             $this->providerService,
             $this->contactService,
-            $this->session,
+            $this->requestStack,
             $this->menuService,
             $this->authenticationRepository,
             $this->sessionRepository,
@@ -266,7 +292,7 @@ class AuthenticateTest extends TestCase
             $this->authenticationService,
             $this->providerService,
             $this->contactService,
-            $this->session,
+            $this->requestStack,
             $this->menuService,
             $this->authenticationRepository,
             $this->sessionRepository,
@@ -323,11 +349,6 @@ class AuthenticateTest extends TestCase
             ->method('addUser')
             ->with($this->contact);
 
-        $this->session
-            ->expects($this->once())
-            ->method('getId')
-            ->willReturn('abdef');
-
         $this->authenticationService
             ->expects($this->once())
             ->method('findAuthenticationTokensByToken')
@@ -338,7 +359,7 @@ class AuthenticateTest extends TestCase
             $this->authenticationService,
             $this->providerService,
             $this->contactService,
-            $this->session,
+            $this->requestStack,
             $this->menuService,
             $this->authenticationRepository,
             $this->sessionRepository,
@@ -392,7 +413,7 @@ class AuthenticateTest extends TestCase
             $this->authenticationService,
             $this->providerService,
             $this->contactService,
-            $this->session,
+            $this->requestStack,
             $this->menuService,
             $this->authenticationRepository,
             $this->sessionRepository,
@@ -444,11 +465,6 @@ class AuthenticateTest extends TestCase
             ->method('updateUser')
             ->with($this->contact);
 
-        $this->session
-            ->expects($this->once())
-            ->method('getId')
-            ->willReturn('abdef');
-
         $this->authenticationService
             ->expects($this->once())
             ->method('findAuthenticationTokensByToken')
@@ -459,7 +475,7 @@ class AuthenticateTest extends TestCase
             $this->authenticationService,
             $this->providerService,
             $this->contactService,
-            $this->session,
+            $this->requestStack,
             $this->menuService,
             $this->authenticationRepository,
             $this->sessionRepository,
@@ -508,11 +524,6 @@ class AuthenticateTest extends TestCase
             ->method('updateUser')
             ->with($this->contact);
 
-        $this->session
-            ->expects($this->any())
-            ->method('getId')
-            ->willReturn('abdef');
-
         $this->authenticationService
             ->expects($this->once())
             ->method('findAuthenticationTokensByToken')
@@ -542,7 +553,7 @@ class AuthenticateTest extends TestCase
             $this->authenticationService,
             $this->providerService,
             $this->contactService,
-            $this->session,
+            $this->requestStack,
             $this->menuService,
             $this->authenticationRepository,
             $this->sessionRepository,
@@ -586,11 +597,6 @@ class AuthenticateTest extends TestCase
             ->method('exists')
             ->willReturn(true);
 
-        $this->session
-            ->expects($this->any())
-            ->method('getId')
-            ->willReturn('abdef');
-
         $this->authenticationService
             ->expects($this->once())
             ->method('findAuthenticationTokensByToken')
@@ -611,7 +617,7 @@ class AuthenticateTest extends TestCase
             $this->authenticationService,
             $this->providerService,
             $this->contactService,
-            $this->session,
+            $this->requestStack,
             $this->menuService,
             $this->authenticationRepository,
             $this->sessionRepository,
@@ -658,11 +664,6 @@ class AuthenticateTest extends TestCase
             ->method('exists')
             ->willReturn(true);
 
-        $this->session
-            ->expects($this->any())
-            ->method('getId')
-            ->willReturn('abdef');
-
         $this->authenticationService
             ->expects($this->once())
             ->method('findAuthenticationTokensByToken')
@@ -683,7 +684,7 @@ class AuthenticateTest extends TestCase
             $this->authenticationService,
             $this->providerService,
             $this->contactService,
-            $this->session,
+            $this->requestStack,
             $this->menuService,
             $this->authenticationRepository,
             $this->sessionRepository,
