@@ -53,10 +53,6 @@ class MonitoringServiceTest extends TestCase
      * @var HostConfigurationServiceInterface&\PHPUnit\Framework\MockObject\MockObject
      */
     private $hostConfiguration;
-    /**
-     * @var MonitoringServerServiceInterface&\PHPUnit\Framework\MockObject\MockObject
-     */
-    private $monitoringServerService;
 
     protected function setUp(): void
     {
@@ -64,13 +60,12 @@ class MonitoringServiceTest extends TestCase
         $this->accessGroupRepository = $this->createMock(AccessGroupRepositoryInterface::class);
         $this->serviceConfiguration = $this->createMock(ServiceConfigurationServiceInterface::class);
         $this->hostConfiguration = $this->createMock(HostConfigurationServiceInterface::class);
-        $this->monitoringServerService = $this->createMock(MonitoringServerServiceInterface::class);
     }
 
     /**
      * @throws \Exception
      */
-    public function testFindServices()
+    public function testFindServices(): void
     {
         $service = (new Service())
             ->setId(1)
@@ -84,8 +79,7 @@ class MonitoringServiceTest extends TestCase
             $this->monitoringRepository,
             $this->accessGroupRepository,
             $this->serviceConfiguration,
-            $this->hostConfiguration,
-            $this->monitoringServerService
+            $this->hostConfiguration
         );
 
         $servicesFound = $monitoringService->findServices();
@@ -99,7 +93,7 @@ class MonitoringServiceTest extends TestCase
     /**
      * @throws \Exception
      */
-    public function testFindServicesByHost()
+    public function testFindServicesByHost(): void
     {
         $service = (new Service())
             ->setId(1)
@@ -115,8 +109,7 @@ class MonitoringServiceTest extends TestCase
             $this->monitoringRepository,
             $this->accessGroupRepository,
             $this->serviceConfiguration,
-            $this->hostConfiguration,
-            $this->monitoringServerService
+            $this->hostConfiguration
         );
 
         $servicesFound = $monitoringService->findServicesByHost($hostId);
@@ -130,7 +123,7 @@ class MonitoringServiceTest extends TestCase
     /**
      * @throws \Exception
      */
-    public function testFindHosts()
+    public function testFindHosts(): void
     {
         $service = (new Service())
             ->setId(1)
@@ -153,8 +146,7 @@ class MonitoringServiceTest extends TestCase
             $this->monitoringRepository,
             $this->accessGroupRepository,
             $this->serviceConfiguration,
-            $this->hostConfiguration,
-            $this->monitoringServerService
+            $this->hostConfiguration
         );
 
         /**
@@ -181,7 +173,7 @@ class MonitoringServiceTest extends TestCase
     /**
      * @throws \Exception
      */
-    public function testFindServiceGroups()
+    public function testFindServiceGroups(): void
     {
         $service = (new Service())
             ->setId(1)
@@ -204,8 +196,7 @@ class MonitoringServiceTest extends TestCase
             $this->monitoringRepository,
             $this->accessGroupRepository,
             $this->serviceConfiguration,
-            $this->hostConfiguration,
-            $this->monitoringServerService
+            $this->hostConfiguration
         );
 
         /**
@@ -225,7 +216,7 @@ class MonitoringServiceTest extends TestCase
     /**
      * @throws \Exception
      */
-    public function testFindOneService()
+    public function testFindOneService(): void
     {
         $service = (new Service())
             ->setId(1)
@@ -244,8 +235,7 @@ class MonitoringServiceTest extends TestCase
             $this->monitoringRepository,
             $this->accessGroupRepository,
             $this->serviceConfiguration,
-            $this->hostConfiguration,
-            $this->monitoringServerService
+            $this->hostConfiguration
         );
 
         $oneService = $monitoringService->findOneService($host->getId(), $service->getId());
@@ -260,7 +250,7 @@ class MonitoringServiceTest extends TestCase
     /**
      * @throws \Exception
      */
-    public function testFindOneHost()
+    public function testFindOneHost(): void
     {
         $service = (new Service())
             ->setId(1)
@@ -280,8 +270,7 @@ class MonitoringServiceTest extends TestCase
             $this->monitoringRepository,
             $this->accessGroupRepository,
             $this->serviceConfiguration,
-            $this->hostConfiguration,
-            $this->monitoringServerService
+            $this->hostConfiguration
         );
 
         $hostFound = $monitoringService->findOneHost($host->getId());
@@ -296,7 +285,7 @@ class MonitoringServiceTest extends TestCase
     /**
      * @throws \Exception
      */
-    public function testFindHostGroups()
+    public function testFindHostGroups(): void
     {
         $service = (new Service())
             ->setId(3)
@@ -319,8 +308,7 @@ class MonitoringServiceTest extends TestCase
             $this->monitoringRepository,
             $this->accessGroupRepository,
             $this->serviceConfiguration,
-            $this->hostConfiguration,
-            $this->monitoringServerService
+            $this->hostConfiguration
         );
 
         /**
@@ -340,7 +328,7 @@ class MonitoringServiceTest extends TestCase
     /**
      * @throws \Exception
      */
-    public function testIsHostExist()
+    public function testIsHostExist(): void
     {
         $host = (new Host())
             ->setId(1)
@@ -355,8 +343,7 @@ class MonitoringServiceTest extends TestCase
             $this->monitoringRepository,
             $this->accessGroupRepository,
             $this->serviceConfiguration,
-            $this->hostConfiguration,
-            $this->monitoringServerService
+            $this->hostConfiguration
         );
 
         // First test when the 'findOneHost' returns one host
@@ -371,7 +358,7 @@ class MonitoringServiceTest extends TestCase
     /**
      * @throws \Exception
      */
-    public function testIsServiceExist()
+    public function testIsServiceExist(): void
     {
         $host = (new Host())
             ->setId(1)
@@ -390,8 +377,7 @@ class MonitoringServiceTest extends TestCase
             $this->monitoringRepository,
             $this->accessGroupRepository,
             $this->serviceConfiguration,
-            $this->hostConfiguration,
-            $this->monitoringServerService
+            $this->hostConfiguration
         );
 
         $exists = $monitoringService->isServiceExists($host->getId(), $service->getId());
@@ -401,7 +387,7 @@ class MonitoringServiceTest extends TestCase
     /**
      * @throws \Exception
      */
-    public function testFindServiceGroupsByHostAndService()
+    public function testFindServiceGroupsByHostAndService(): void
     {
         $service = (new Service())
             ->setId(1)
@@ -425,8 +411,7 @@ class MonitoringServiceTest extends TestCase
             $this->monitoringRepository,
             $this->accessGroupRepository,
             $this->serviceConfiguration,
-            $this->hostConfiguration,
-            $this->monitoringServerService
+            $this->hostConfiguration
         );
 
         /**
