@@ -16,18 +16,18 @@ import queryString from 'query-string';
 import FullscreenIcon from '@material-ui/icons/Fullscreen';
 import { withStyles, createStyles, Fab } from '@material-ui/core';
 
-import { ThemeProvider } from '@centreon/ui';
+import { LoadingSkeleton, ThemeProvider } from '@centreon/ui';
 
-import Header from './Header';
 import { history } from './store';
-import Nagigation from './Navigation';
-import Footer from './Footer';
 import axios from './axios';
 import { fetchExternalComponents } from './redux/actions/externalComponentsActions';
 import PageLoader from './components/PageLoader';
 import Provider from './Provider';
 
 const MainRouter = React.lazy(() => import('./components/mainRouter'));
+const Header = React.lazy(() => import('./Header'));
+const Navigation = React.lazy(() => import('./Navigation'));
+const Footer = React.lazy(() => import('./Footer'));
 
 const styles = createStyles({
   content: {
@@ -159,9 +159,21 @@ class App extends Component<Props, State> {
         <ConnectedRouter history={history}>
           <ThemeProvider>
             <div className={classes.wrapper}>
-              {!min && <Nagigation />}
+              {!min && (
+                <React.Suspense
+                  fallback={<LoadingSkeleton height="100%" width={45} />}
+                >
+                  <Navigation />
+                </React.Suspense>
+              )}
               <div className={classes.content} id="content">
-                {!min && <Header />}
+                {!min && (
+                  <React.Suspense
+                    fallback={<LoadingSkeleton height={56} width="100%" />}
+                  >
+                    <Header />
+                  </React.Suspense>
+                )}
                 <div
                   className={classes.fullScreenWrapper}
                   id="fullscreen-wrapper"
@@ -178,7 +190,13 @@ class App extends Component<Props, State> {
                     </div>
                   </Fullscreen>
                 </div>
-                {!min && <Footer />}
+                {!min && (
+                  <React.Suspense
+                    fallback={<LoadingSkeleton height={30} width="100%" />}
+                  >
+                    <Footer />
+                  </React.Suspense>
+                )}
               </div>
               <Fab
                 className={classes.fullscreenButton}
