@@ -19,7 +19,7 @@ import {
   SeverityCode,
   StatusCounter,
 } from '@centreon/ui';
-import { userAtom } from '@centreon/centreon-frontend/packages/ui-context/src';
+import { userAtom } from '@centreon/ui-context';
 
 import styles from '../../header.scss';
 import {
