@@ -45,6 +45,8 @@ const retrievedDefaultParameters = {
   monitoring_default_acknowledgement_persistent: true,
   monitoring_default_acknowledgement_sticky: false,
   monitoring_default_downtime_duration: 1458,
+  monitoring_default_downtime_fixed: true,
+  monitoring_default_downtime_with_services: false,
   monitoring_default_refresh_interval: 15,
 };
 
@@ -159,8 +161,10 @@ describe(Provider, () => {
       expect(atomsValue.downtime).toEqual({
         default_duration:
           retrievedDefaultParameters.monitoring_default_downtime_duration,
-        default_fixed: false,
-        default_with_services: false,
+        default_fixed:
+          retrievedDefaultParameters.monitoring_default_downtime_fixed,
+        default_with_services:
+          retrievedDefaultParameters.monitoring_default_downtime_with_services,
       });
       expect(atomsValue.refreshInterval).toEqual(
         retrievedDefaultParameters.monitoring_default_refresh_interval,
