@@ -2,8 +2,12 @@ import * as React from 'react';
 
 import { useTheme, Grid } from '@material-ui/core';
 
-import ResourceActions from './Resource';
-import GlobalActions, { Props } from './Refresh';
+import { Props } from './Refresh';
+import GlobalActionsSkeleton from './GlobalActionsSkeleton';
+import ResourceActionsSkeleton from './ResourceActionsSkeleton';
+
+const ResourceActions = React.lazy(() => import('./Resource'));
+const GlobalActions = React.lazy(() => import('./Refresh'));
 
 const Actions = ({ onRefresh }: Props): JSX.Element => {
   const theme = useTheme();
@@ -11,10 +15,14 @@ const Actions = ({ onRefresh }: Props): JSX.Element => {
   return (
     <Grid container>
       <Grid item>
-        <ResourceActions />
+        <React.Suspense fallback={<ResourceActionsSkeleton />}>
+          <ResourceActions />
+        </React.Suspense>
       </Grid>
       <Grid item style={{ paddingLeft: theme.spacing(3) }}>
-        <GlobalActions onRefresh={onRefresh} />
+        <React.Suspense fallback={<GlobalActionsSkeleton />}>
+          <GlobalActions onRefresh={onRefresh} />
+        </React.Suspense>
       </Grid>
     </Grid>
   );
