@@ -1,3 +1,5 @@
+import * as React from 'react';
+
 import { ResourceEndpoints } from '../../models';
 
 import { TabProps } from '.';
@@ -7,7 +9,7 @@ export type TabEndpoints = Omit<ResourceEndpoints, 'details'>;
 export type TabId = 0 | 1 | 2 | 3 | 4;
 
 export interface Tab {
-  Component: (props: TabProps) => JSX.Element;
+  Component: React.LazyExoticComponent<(props: TabProps) => JSX.Element>;
   getIsActive: (details) => boolean;
   id: TabId;
   title: string;
