@@ -19,12 +19,12 @@
  *
  */
 
+
 include_once __DIR__ . "/../../class/centreonLog.class.php";
 $centreonLog = new CentreonLog();
 
+//error specific content
 $versionOfTheUpgrade = 'UPGRADE - 22.04.0-beta.1: ';
-
-$pearDB = new CentreonDB();
 
 try {
     $errorMessage = "Unable to create table 'password_security_policy'";
@@ -89,7 +89,6 @@ try {
     $pearDB->query("ALTER TABLE `contact` DROP COLUMN `contact_passwd`");
 
     $errorMessage = 'Impossible to add "contact_js_effects" column to "contact" table';
-
     if (!$pearDB->isColumnExist('contact', 'contact_js_effects')) {
         $pearDB->query(
             "ALTER TABLE `contact`
@@ -97,6 +96,16 @@ try {
             AFTER `contact_comment`"
         );
     }
+
+    $errorMessage = 'Unable to update the description in cb_field';
+    $statement = $pearDB->query("
+        UPDATE cb_field
+        SET `description` = 'Time in seconds to wait between each connection attempt (Default value: 30s).'
+        WHERE `cb_field_id` = 31
+    ");
+
+    $errorMessage = 'Unable to delete logger entry in cb_tag';
+    $statement = $pearDB->query("DELETE FROM cb_tag WHERE tagname = 'logger'");
 } catch (\Exception $e) {
     if ($pearDB->inTransaction()) {
         $pearDB->rollBack();
